fix(Overview): cleanup
